+++ azd/src/main/java/org/azd/interfaces/GraphDetails.java
@@ -1,10 +1,7 @@
 package org.azd.interfaces;
 
 import org.azd.exceptions.AzDException;
-import org.azd.graph.types.GraphGroup;
-import org.azd.graph.types.GraphGroups;
-import org.azd.graph.types.GraphUser;
-import org.azd.graph.types.GraphUsers;
+import org.azd.graph.types.*;
 
 public interface GraphDetails {
     GraphUser createUser(String emailId, String userDescriptor) throws AzDException;
@@ -15,4 +12,13 @@ public interface GraphDetails {
     GraphUsers getUsers(String continuationToken, String scopeDescriptor, String subjectTypes) throws AzDException;
     GraphGroup getGroup(String groupDescriptor) throws AzDException;
     GraphGroups getGroups() throws AzDException;
+    GraphMemberships getGroupMembersOf(String groupDescriptor) throws AzDException;
+    GraphMemberships getMemberOfGroups(String subjectDescriptor) throws AzDException;
+    GraphMembership addMembership(String subjectDescriptor, String groupDescriptor) throws AzDException;
+    Void removeMembership(String subjectDescriptor, String groupDescriptor) throws AzDException;
+    GraphMembership createGroup(String displayName, String description) throws AzDException;
+    GraphMembership createGroup(String displayName, String description, String projectDescriptor) throws AzDException;
+    Void deleteGroup(String groupDescriptor) throws AzDException;
+    GraphDescriptor getDescriptor(String storageKey) throws AzDException;
+    SubjectLookupResponse subjectLookup(String... descriptors) throws AzDException;
 }